+++ src/dms_provider_bridge/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ["__version__"]
 
-__version__ = "0.7.8-beta"
+__version__ = "0.7.9-beta"